+++ NEWS.md
@@ -1,3 +1,15 @@
+# Development Version
+
+* New Functions:
+    * `tk_make_future_timeseries` creates a future time-based index from an existing time-based index. 
+    * `tk_get_timeseries_signature` returns a tibble with the time series index decomposed into year, quarter, month, etc.
+    * `tk_get_timeseries_summary` returns summary metrics for the time series index including number of observations, start, end, units, scale, diff summary (summary for frequency in seconds), etc.
+    * `tk_augment_timeseries_signature` adds the time series signature to a `tbl` (with time base), `xts` or `zoo` object.
+    * `tk_get_timeseries_variables` returns the variable column names for date, datetime, yearmon, or yearqtr variables in a data frame. 
+* Bug Fixes:
+    * `tk_index`: Fixed timezone issue with yearmon and yearqtr classes. 
+    
+
 
 # timekit 0.1.0